Added initial version of schematic/board files
Temporary layout for parts as an "extension" board for connection to an
arduino or other dev kit. Does not represent a full, functional device,
but just as a planning exercise. Will likely have the final circuit on
perfboard with a bunch of auxilaries connected, since this will likely
only be used a few times and doesn't need to be a polished device.