Added Rev 0 of Arduino firmware.
Functioning initial implementation of the Arduino firmware (used on
Arduino Uno), may update and clean up a few things, but it's mostly in
its final form. Will likely be using a Teensy LC or smaller Arduino
board, so may need to update some of the pin numbering. Corresponding
update to schematic will be made when final board is decided on.